src > views > SchoolApp > Students > SmartList > SmartListDetail > index
The smart list is one of many important features I built that allows Administrators to perform an action on their student(s). The smart list includes a form for filtering the data, a form to enable/disable the columns to be displayed, a grid to display all/filtered students, and buttons that allow the user to save/load a report, message the selected students, or export the results to a pdf/excel file.